I have no idea what I'm doing
Hey, you know what would be cool? A tower defense game with civ
components in it that runs in the browser. I should make that. Yeah.  I
don’t need to have done anything like it before I should just do it,
I’ll figure it out as I go along. Here something, I don’t know what it
is or how it resembles the idea, but here is some code with things that
move around.