Formatting makes this commit impossible to stage
Reformatting made this commit made this more work to stage then  I want
to put in for a personal project.
Mouse event queues are in in some form or another.
There is a simple collision detection in Rect objects.
You can select and move Attack entities. I need to move that to a
different component, maybe movable?